feat: add shared utilities and enhance WYSIWYG Studio integration; implement JSON parsing and string transformation functions
