+++ .vscode/settings.json
@@ -1,6 +1,7 @@
 {
   "cSpell.words": [
     "Avenir",
+    "CMSWYSIWYG",
     "compkeys",
     "grapesjs",
     "grayscale",
